Module divide
모듈 분리 
Schema 목록 가져오기 
Data 가져오기 
UrlComponent : class -> url path binding
