+++ Dockerfile
@@ -4,16 +4,16 @@ FROM java:8
 LABEL maintainer="minssan9@gmail.com"
 
 # Add a volume to /tmp
-VOLUME /charger
+VOLUME /fn
 
 # Make port 8080 available to the world outside this container
 EXPOSE 54000
 
 # The application's jar file
-ARG JAR_FILE=build/libs/charger-*.jar
+ARG JAR_FILE=build/libs/fn-*.jar
 
 # Add the application's jar to the container
-ADD ${JAR_FILE} charger-springboot.jar
+ADD ${JAR_FILE} fn-springboot.jar
 
 # Run the jar file
-ENTRYPOINT ["java", "-jar","/charger-springboot.jar"]
+ENTRYPOINT ["java", "-jar","/fn-springboot.jar"]
