+++ batch/build.gradle
@@ -1,0 +1,29 @@
+plugins {
+    id 'org.springframework.boot'
+}
+
+group = 'com.voyagerss'
+version = '0.1.0'
+
+repositories {
+    mavenCentral()
+}
+
+
+bootJar {
+    manifest {
+        attributes 'Start-Class': 'com.service.BatchApplication'
+    }
+}
+dependencies {
+    compile project(':core')
+    compile project(':service')
+
+    compile 'org.springframework.boot:spring-boot-starter-batch'
+
+    testCompile('org.springframework.batch:spring-batch-test')
+}
+
+test {
+    useJUnitPlatform()
+}
