+++ batch/src/main/java/com/batch/BatchApplication.java
@@ -1,0 +1,21 @@
+package com.batch;
+
+import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+@EnableBatchProcessing
+@EnableScheduling
+@EnableJpaRepositories(basePackages = {"com.service.repository"})
+@EntityScan(basePackages = {"com.service.domain"})
+@SpringBootApplication
+public class BatchApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(BatchApplication.class, args);
+    }
+
+}
