+++ batch/src/main/java/com/batch/tasklet/EcosJobStep2Tasklet.java
@@ -1,0 +1,28 @@
+package com.scheduler.tasklet;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.batch.core.StepContribution;
+import org.springframework.batch.core.configuration.annotation.StepScope;
+import org.springframework.batch.core.scope.context.ChunkContext;
+import org.springframework.batch.core.step.tasklet.Tasklet;
+import org.springframework.batch.repeat.RepeatStatus;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+@Slf4j
+@Component
+@StepScope
+public class SampleJobStep2Tasklet implements Tasklet {
+
+    @Value("#{jobParameters[str]}")
+    private String str;
+
+    @Override
+    public RepeatStatus execute(StepContribution contribution, ChunkContext chunkContext) throws Exception {
+        log.info("sampleStep2 start : {}", str);
+
+        // step2 logic
+
+        return RepeatStatus.FINISHED;
+    }
+}
