+++ batch/src/test/java/com/batch/BatchApplicationTests.java
@@ -1,0 +1,13 @@
+package com.batch;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BatchApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
