+++ build.gradle
@@ -6,6 +6,7 @@ group  'com.minssan9'
 
 allprojects {
 
+
 }
 
 subprojects {
@@ -37,7 +38,10 @@ subprojects {
         annotationProcessor 'org.projectlombok:lombok'
 
         testImplementation 'org.springframework.boot:spring-boot-starter-test'
+
+        annotationProcessor 'org.springframework.boot:spring-boot-configuration-processor'
     }
+
 
     test {
         useJUnitPlatform()
