+++ core/src/test/java/com/core/CoreApplicationTests.java
@@ -1,0 +1,13 @@
+package com.core;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CoreApplicationTests {
+
+//    @Test
+    void contextLoads() {
+    }
+
+}
