--- service/src/main/java/com/financial/service/config/KrBankProperties.java
@@ -1,12 +1,0 @@
-package com.financial.service.config;
-
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-@Component
-@ConfigurationProperties("krbank")
-@Data
-public class KrBankProperties {
-    public  String apikey;
-}
