--- service/src/main/java/com/financial/service/repository/KrBankDataRepository.java
@@ -1,11 +1,0 @@
-package com.financial.service.repository;
-
-import com.financial.service.domain.KrBankData;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public
-interface KrBankDataRepository extends JpaRepository<KrBankData, Long>{
-
-}
