+++ service/src/main/java/com/service/config/KrBankProperties.java
@@ -1,0 +1,12 @@
+package com.service.config;
+
+import lombok.Data;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+@Component
+@ConfigurationProperties("krbank")
+@Data
+public class KrBankProperties {
+    public  String apikey;
+}
