+++ service/src/main/java/com/service/dto/StatisticSearch.java
@@ -1,0 +1,2 @@
+package com.service.dto;
+
