+++ service/src/main/java/com/service/repository/KrBankDataRepository.java
@@ -1,0 +1,11 @@
+package com.service.repository;
+
+import com.service.domain.KrBankData;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public
+interface KrBankDataRepository extends JpaRepository<KrBankData, Long>{
+
+}
