+++ service/src/main/java/com/service/repository/KrBankSchemaRepository.java
@@ -1,0 +1,11 @@
+package com.service.repository;
+
+import com.service.domain.KrBankSchema;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public
+interface KrBankSchemaRepository extends JpaRepository<KrBankSchema, Long> {
+
+}
