+++ service/src/main/java/com/service/service/KrBankService.java
@@ -1,0 +1,15 @@
+package com.service.service;
+
+import com.service.domain.KrBankData;
+import com.service.domain.KrBankSchema;
+import com.service.dto.KrBankRequest;
+import java.util.List;
+import org.springframework.stereotype.Service;
+
+
+//@Slf4j
+@Service
+interface KrBankService {
+    List<KrBankSchema> getSchema(KrBankRequest krBankRequest );
+    List<KrBankData> getData(KrBankRequest krBankRequest );
+}
