+++ service/src/test/java/com/service/ServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.service;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
